requirementsPython.txt
requests==2.22.0
opencv-python==4.1.2.30
Pillow==6.2.1

